+++ test/models/animal.rb
@@ -4,6 +4,5 @@ class Animal
     text_start: [:name],
     suggest: [:name],
     index_name: -> { "#{name.tableize}-#{Date.today.year}#{Searchkick.index_suffix}" },
-    callbacks: :async,
-    wordnet: ENV["WORDNET"]
+    callbacks: :async
 end