+++ test/test_helper.rb
@@ -25,9 +25,6 @@ end
 Searchkick.search_timeout = 5
 Searchkick.index_suffix = ENV["TEST_ENV_NUMBER"] # for parallel tests
 
-# add to elasticsearch-7.0.0/config/
-Searchkick.wordnet_path = "wn_s.pl" if ENV["WORDNET"]
-
 puts "Running against #{Searchkick.opensearch? ? "OpenSearch" : "Elasticsearch"} #{Searchkick.server_version}"
 
 Searchkick.redis =